# [PERF] evaluation: use custom data structures instead of Map and Set