--- a/src/plugins/ui_core_views/cell_evaluation/formula_dependency_graph.ts
+++ b/src/plugins/ui_core_views/cell_evaluation/formula_dependency_graph.ts
@@ -1,5 +1,7 @@
-import { JetSet } from "../../../helpers";
-import { PositionBitsEncoder, PositionId } from "./evaluator";
+import { positions, positionToZone } from "../../../helpers";
+import { CellPosition } from "../../../types";
+import { PositionMap } from "./position_map";
+import { PositionSet } from "./position_set";
 import { RTreeBoundingBox, RTreeItem, SpreadsheetRTree } from "./r_tree";
 
 /**
@@ -10,27 +12,30 @@ import { RTreeBoundingBox, RTreeItem, SpreadsheetRTree } from "./r_tree";
  * It uses an R-Tree data structure to efficiently find dependent cells.
  */
 export class FormulaDependencyGraph {
-  private readonly dependencies: Map<PositionId, RTreeItem<PositionId>[]> = new Map();
-  private readonly rTree: SpreadsheetRTree<PositionId>;
+  private readonly dependencies: PositionMap<RTreeItem<CellPosition>[]> = new PositionMap();
+  private readonly rTree: SpreadsheetRTree<CellPosition>;
 
-  constructor(private readonly encoder: PositionBitsEncoder, data: RTreeItem<PositionId>[] = []) {
+  constructor(
+    private readonly createEmptyPositionSet: () => PositionSet,
+    data: RTreeItem<CellPosition>[] = []
+  ) {
     this.rTree = new SpreadsheetRTree(data);
   }
 
-  removeAllDependencies(formulaPositionId: PositionId) {
-    const ranges = this.dependencies.get(formulaPositionId);
+  removeAllDependencies(formulaPosition: CellPosition) {
+    const ranges = this.dependencies.get(formulaPosition);
     if (!ranges) {
       return;
     }
     for (const range of ranges) {
       this.rTree.remove(range);
     }
-    this.dependencies.delete(formulaPositionId);
+    this.dependencies.delete(formulaPosition);
   }
 
-  addDependencies(formulaPositionId: PositionId, dependencies: RTreeBoundingBox[]): void {
+  addDependencies(formulaPosition: CellPosition, dependencies: RTreeBoundingBox[]): void {
     const rTreeItems = dependencies.map(({ sheetId, zone }) => ({
-      data: formulaPositionId,
+      data: formulaPosition,
       boundingBox: {
         zone,
         sheetId,
@@ -39,11 +44,11 @@ export class FormulaDependencyGraph {
     for (const item of rTreeItems) {
       this.rTree.insert(item);
     }
-    const existingDependencies = this.dependencies.get(formulaPositionId);
+    const existingDependencies = this.dependencies.get(formulaPosition);
     if (existingDependencies) {
       existingDependencies.push(...rTreeItems);
     } else {
-      this.dependencies.set(formulaPositionId, rTreeItems);
+      this.dependencies.set(formulaPosition, rTreeItems);
     }
   }
 
@@ -52,22 +57,27 @@ export class FormulaDependencyGraph {
    * in the correct order they should be evaluated.
    * This is called a topological ordering (excluding cycles)
    */
-  getCellsDependingOn(ranges: RTreeBoundingBox[]): Set<PositionId> {
-    const visited: JetSet<PositionId> = new JetSet<PositionId>();
+  getCellsDependingOn(ranges: RTreeBoundingBox[]): PositionSet {
+    const visited = this.createEmptyPositionSet();
     const queue: RTreeBoundingBox[] = Array.from(ranges).reverse();
-
     while (queue.length > 0) {
       const range = queue.pop()!;
-      visited.addMany(this.encoder.encodeBoundingBox(range));
+      visited.addMany(
+        positions(range.zone).map((position) => ({ sheetId: range.sheetId, ...position }))
+      );
 
-      const impactedPositionIds = this.rTree.search(range).map((dep) => dep.data);
-      for (const positionId of impactedPositionIds) {
-        if (!visited.has(positionId)) {
-          queue.push(this.encoder.decodeToBoundingBox(positionId));
+      const impactedPositions = this.rTree.search(range).map((dep) => dep.data);
+      for (const position of impactedPositions) {
+        if (!visited.has(position)) {
+          queue.push({ sheetId: position.sheetId, zone: positionToZone(position) });
         }
       }
     }
-    visited.deleteMany(ranges.flatMap((r) => this.encoder.encodeBoundingBox(r)));
+    visited.deleteMany(
+      ranges.flatMap((r) =>
+        positions(r.zone).map((position) => ({ sheetId: r.sheetId, ...position }))
+      )
+    );
     return visited;
   }
 }